Implement FlatArray EqualityComparer nested
