--- src/collections-flat-array/Collections.FlatArray/FlatArray.T/FlatArray.InternalItemsSame.cs
@@ -1,11 +1,0 @@
-﻿using System.Runtime.CompilerServices;
-
-namespace System.Collections.Generic;
-
-partial class FlatArray<T>
-{
-    [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    internal static bool InternalItemsSame(FlatArray<T> left, FlatArray<T> right)
-        =>
-        ReferenceEquals(left.items, right.items);
-}
